Fixed: systemd expects /srv/data directory
We added a ReadWriteDir allowance in SystemD for /srv/data.  However, not all applications use it,
so we left it out of the Tomcat playbook.  The problem is that the ReadWriteDir declaration causes a
crash if the directory does not exists.  So we must create the directory during Tomcat installation,
even if a webapp does not need the data directory.